[GUI][ETL] Avoid packaging unused dependencies
