+++ etl/build.gradle
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 plugins {
-    // Apply the scala Plugin to add support for Scala.
+    id "com.github.johnrengelman.shadow" version "7.1.2"
     id 'scala'
 }
 
@@ -34,10 +34,11 @@ def versions = [
 ]
 
 dependencies {
-    implementation "org.scala-lang:scala-library:${versions["scala"]}"
+    // Astraea etl is run by spark framework, so we don't need to package spark-related dependencies
+    compileOnly "org.scala-lang:scala-library:${versions["scala"]}"
+    compileOnly "org.apache.spark:spark-sql_2.12:${versions["spark_sql"]}"
+    compileOnly "org.apache.spark:spark-sql-kafka-0-10_2.12:${versions["spark_kafka"]}"
     implementation "org.apache.kafka:kafka-clients:${versions["kafka"]}"
-    implementation "org.apache.spark:spark-sql_2.12:${versions["spark_sql"]}"
-    implementation "org.apache.spark:spark-sql-kafka-0-10_2.12:${versions["spark_kafka"]}"
     implementation project(':common')
 
     testImplementation "org.junit.jupiter:junit-jupiter:${versions["junit"]}"
